+++ src/sensor_interface.cpp
@@ -6,9 +6,9 @@ inline void initializeIMU() {
 inline void initializeDepthSensor() {
 }
 
-inline void updateIMUReadings(int16_t& ax, int16_t& ay, int16_t& az, int16_t& gx, int16_t& gy, int16_t& gz, int16_t& mx, int16_t& my, int16_t& mz) {
+inline void updateIMUReadings(float& ax, float& ay, float& az, float& gx, float& gy, float& gz, float& mx, float& my, float& mz) {
 }
 
-inline void updateDepthSensorReadings(int16_t& depth) {
+inline void updateDepthSensorReadings(float& depth) {
 }
